몽고DB Pagenation 적용

+++ src/main/kotlin/org/inner/circle/o2oserver/store/domain/store/StoreInfo.kt
@@ -1,6 +1,6 @@
 package org.inner.circle.o2oserver.store.domain.store
 
 class StoreListInfo(
-    val totalCount: Int,
+    val totalCount: Long,
     val stores: List<Store>,
 )
